requirements.txt
gradio==4.19.2
# fastapi==0.110.0
qdrant-client==1.7.3
langchain==0.1.11
# sentence-transformers==2.5.1
pdfminer.six==20231228
openai==1.13.3
python-dotenv==1.0.1
pyyaml==6.0.1
tiktoken==0.6.0
langchain-openai==0.0.8